fix: Remove pricing adapter import cycle
